tree.h: format argument lists of read_tree() users
Re-indent and format the argument list of read_tree() callbacks, and
the relevant functions they call.

This is a whitespace-only change to make changes in a later series
easier to read. In that series I'll be adding a new argument on the
"mode" line, so leave space on it for a new argument.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>