object-file.c: take type id, not string, in read_object_with_reference()
Make the read_object_with_reference() function take "enum object_type"
instead of a "const char *" with a type name that it converted via
type_from_string().

Out of the nine callers of this function, only one wanted to pass a
"const char *". The others were simply passing along the
{commit,tree}_type string constants.

This API design dates back to f4913f91a96 ([PATCH] Accept commit in
some places when tree is needed., 2005-04-20), at that time there
wasn't an API like type_from_string(). That only arrived in
df8436622fb (formalize typename(), and add its reverse
type_from_string(), 2007-02-26).

Since the introduction on of the function it's become clear that this
is the more common way to use this API. So let's change it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>