hash-object: parse the "-t" option into "enum object_type"
Change the "hash-object -t <type>" option so it's parsed directly into
an "enum object type", and validated with
type_from_string_gently(). Previously if e.g. "garbage" was supplied
as a type we'd make it all the way into hash_fd() and friends before
we'd die. Now the parsing and erroring happens closer hash-objects.

The parse_opt_object_name() is mostly copy/pasted from the
parse_opt_object_id() above it, which was added in 7d3488eb893 (rebase
-i: use struct commit when parsing options, 2019-04-17).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>