tree-walk.h API: add a get_tree_entry_path() function
Add a get_tree_entry_path() variant in addition to
get_tree_entry_path_mode(). This is for those callers that don't need
the "mode" filled for them.

There are callers here which don't need the "struct object_id" filled
either, but let's leave that for now. I'm focusing downstream code
that depends on "mode" (or "enum object_type").

The code being modified here was introduced in:

 - shift_tree(): 68faf68938e (A new merge stragety[sic] 'subtree'.,
    2007-02-15) for the shift_tree()

 - tree_has_path(): 96e7ffbdc31 (merge-recursive: check for directory
   level conflicts, 2018-04-19)

 - init_notes(): fd53c9eb445 (Speed up git notes lookup, 2009-10-09)

 - diagnose_invalid_oid_path(): 009fee4774d (Detailed diagnosis when
   parsing an object name fails., 2009-12-07)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>