Makefile: generate 'git' as 'cc [...] -o git+ && mv git+ git'
[edit: a better reason for this:
https://www.google.com/search?q=aix+%22The+file+is+in+use+and+cannot+be+overwritten%22&oq=aix+%22The+file+is+in+use+and+cannot+be+overwritten%22&aqs=chrome..69i57.1193j0j7&sourceid=chrome&ie=UTF-8
https://issues.rudder.io/issues/5815
]

Change the compilation of the main 'git' binary to not have the CC
clobber 'git' in-place. This means that e.g. running the test suite
in-place and recompiling won't fail whatever tests happen to be
running for the duration of the binary being regenerated.

This is not a complete solution, in particular I'm not doing this for
any of the git-* binaries, so it's effectively only for
SKIP_DASHED_BUILT_INS=Y.

I think that's fine for a small best-effort solution to this, even
those who rely on dashed binaries for something are likely mostly
using 'git' for whatever scripts they have running out of their
git.git checkout.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>