object tests: add test for unexpected objects in tags
Fix a blind spot in the tests added in 0616617c7e1 (t: introduce tests
for unexpected object types, 2019-04-09), there were no meaningful
tests for checking how we reported on finding the incorrect object
type in a tag, i.e. one that broke the "type" promise in the tag
header.

[TODO: These tests happen to leak memory, so let's mark them as
!SANITIZE_LEAK. I've got another cooking/in-flight change that marks
these tests as passing with SANITIZE=leak].

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>