test-lib: guard against being run as 'prove t*.sh'
We have t/Makefile targets to properly run the tests with prove,
however it's also just as valid to run them as e.g.:

    prove t[0-9]*.sh

But if they're run as:

    prove t*.sh

The prove tool will optimistically try to parse and execute
test-lib.sh and test-lib-functions.sh as Perl code, currently there
are no bad results of doing this other than a wall of error messages,
but let's be helpful and bail out early in this case.

Note that the shebang being added here doesn't mean anything to a
POSIX OS because the files aren't executable, it's picked up by Perl's
TAP tools. If the shebang wasn't here this would still be considered a
Perl file for the purposes of those heuristics.

I could also use the "Bail out!" keyword here, see
614fe01521 (test-lib: bail out when "-v" used under "prove",
2016-10-22) for prior art. I think it's more gentle to just print a
SKIP notice.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>