test-tool: split up test-tool read-cache
Since the "test-tool read-cache" was originally added back in
1ecb5ff141 (read-cache: add simple performance test, 2013-06-09) it's
been growing all sorts of bells and whistles that aren't very
conducive to performance testing the index, e.g. it learned how to
read config.

Let's split what remains of the "test-tool read-cache" into the two
narrow use-cases it's used for.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>