git-config doc: split uploadpack.txt into two included files
Split the included uploadpack.txt file into a file that discusses
things not to do with pack-objects, and one that does. This is to
eventually selectively include this section in the pack-objects
documentation.

This introduces a new config/*/* directory, the contents of e.g. a
config/uploadpack/*.txt will be that part of the existing
config/uploadpack.txt we'll want to grab for more selective inclusion
in other manual pages.

Aside from the trivial Makefile change there is no change here except
a plain moving of lines between files.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>