add docs: include a CONFIGURATION section
Add a CONFIGURATION section to the documentation for git-add. Since
we'll want to use the same blurb as I added in b6a8d09f6d8 (gc docs:
include the "gc.*" section from "config" in "gc", 2019-04-07) let's
make it a template in preparation for repeating this pattern in other
command documentation.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>